# swagger.yaml
swagger: "2.0"
info:
  description: "a RESTful API that retrieves information from NHTSA and presents it to the user."
  version: "1.0.0"
  title: "nhtsa-restful-api"
  contact:
    email: "sbawankar13@outlook.com"
host: ""
basePath: ""
tags:
- name: "manufacturers"
  description: "To get information about Vehicles from nhtsa open-api"
  externalDocs:
    description: "Find out more about the NHTSA API"
    url: "https://vpic.nhtsa.dot.gov/api/"
schemes:
- "https"
- "http"
paths:
  
  /manufacturers:
    get:
      tags:
      - "manufacturers"
      summary: "Get a list of all manufacturers"
      description: "Json output with total number of manufacturers and their list is provided"
      produces:
      - "application/json"
      responses:
        "200":
          description: "OK"
          schema:
            type: "object"
            items:
              $ref: "#/definitions/manufacturer_list"
  /manufacturers/{name}:
    get:
      tags:
      - "manufacturers"
      summary: "Gives Vehicles given manufacturer name"
      description: "list of all makes is given by providing manufacturer name"
      produces:
      - "application/json"
      parameters:
      - name: "name"
        in: "path"
        description: "manufacturer name"
        required: true
        type: "string"
        example: "honda"
      responses:
        "200":
          description: "OK"
          schema:
            type: "object"
            items:
              $ref: "#/definitions/manufacturer_list"
        "400":
          description: "Bad Request"
          schema:
            type: "object"
            items:
              $ref: "#/definitions/err_object"
  /manufacturers/vin/{vin_number}:
    get:
      tags:
      - "manufacturers"
      summary: "the year, make, and model of a vehicle given its VIN"
      description: "the year, make, and model of a vehicle given its VIN"
      produces:
      - "application/json"
      parameters:
      - name: "vin_number"
        in: "path"
        description: "VIN number"
        required: true
        type: "string"
        example: "3N1AB6AP7BL729215"
      responses:
        "200":
          description: "OK"
          schema:
            type: "object"
            items:
              $ref: "#/definitions/vin_details"
        "400":
          description: "Bad Request"
          schema:
            type: "object"
            items:
              $ref: "#/definitions/err_object"
  /manufacturers/Validate_vin/{vin}:
    get:
      tags:
      - "manufacturers"
      summary: "Validate the incoming VIN number"
      description: "Validate the incoming VIN number"
      produces:
      - "application/json"
      parameters:
      - name: "vin"
        in: "path"
        description: "VIN number"
        required: true
        type: "string"
        example: "3N1AB6AP7BL729215"
      responses:
        "200":
          description: "OK"
          schema:
            type: "object"
            items:
              $ref: "#/definitions/success_object"
        "400":
          description: "Bad Request"
          schema:
            type: "object"
            items:
              $ref: "#/definitions/err_object"
definitions:
  manufacturer_list:
    type: "object"
    properties:
      Total Manufacturers:
        type: "string"
        example: "45"
      Manufacturers List:
        type: "string"
        example: "[]"

  vin_details:
    type: "object"
    properties:
      3N1AB6AP7BL729215:
        type: "string"
        example: "[2011 NISSAN Sentra]"

  err_object:
    type: "object"
    properties:
      Result:
        type: "string"
        example: "Bad Request"

  success_object:
    type: "object"
    properties:
      Result:
        type: "string"
        example: "VIN Valid!"